+++ alPyne/utils.py
@@ -1,0 +1,1 @@
+EPSILON = 1.e-10
